requirements.txt
vllm==0.11.0
torch==2.8.0
trl==0.18.1
accelerate~=1.7.0
huggingface-hub~=0.36.0
flashinfer-python~=0.3.1
deepspeed~=0.17.4
datasets~=3.6.0
transformers~=4.55.4
numpy~=2.2.6
pandas~=2.2.3
wandb~=0.20.1
peft~=0.15.2
evaluate~=0.4.6
tqdm~=4.67.1
uvicorn~=0.34.3
fastapi~=0.115.12
httpx~=0.28.1
pydantic~=2.11.6
python-dotenv~=1.1.0
llama-index-core>=0.12.47
llama-index-vector-stores-qdrant~=0.6.1
llama-index-llms-openai-like~=0.4.0
llama-index-question-gen-openai~=0.3.1
qdrant-client~=1.14.3
rouge-score~=0.1.2
matplotlib~=3.9.2